Cleanup ARM builds This is a general cleanup. arm_simd.h was added because we are duplicating too much code in source files. It hit critical mass when we needed the polynomial multiplications in gf2n_sim.cpp
